+++ MoltenVK/MoltenVK/Commands/MVKCommandBuffer.h
@@ -283,8 +283,7 @@ public:
 	void beginRenderpass(MVKCommand* passCmd,
 						 VkSubpassContents subpassContents,
 						 MVKRenderPass* renderPass,
-						 VkExtent2D framebufferExtent,
-						 uint32_t framebufferLayerCount,
+						 MVKFramebuffer* framebuffer,
 						 VkRect2D& renderArea,
 						 MVKArrayRef<VkClearValue> clearValues,
 						 MVKArrayRef<MVKImageView*> attachments);
@@ -306,6 +305,12 @@ public:
 
 	/** Returns the render subpass that is currently active. */
 	MVKRenderSubpass* getSubpass();
+
+	/** The extent of current framebuffer.*/
+	VkExtent2D getFramebufferExtent();
+
+	/** The layer count of current framebuffer.*/
+	uint32_t getFramebufferLayerCount();
 
 	/** Returns the index of the currently active multiview subpass, or zero if the current render pass is not multiview. */
 	uint32_t getMultiviewPassIndex();
@@ -483,12 +488,6 @@ public:
 	/** Indicates whether the current draw is an indexed draw. */
 	bool _isIndexedDraw;
 
-	/** The extent of current framebuffer.*/
-	VkExtent2D _framebufferExtent;
-
-	/** The layer count of current framebuffer.*/
-	uint32_t _framebufferLayerCount;
-
 #pragma mark Construction
 
 	MVKCommandEncoder(MVKCommandBuffer* cmdBuffer);
@@ -513,6 +512,7 @@ protected:
 
 	VkSubpassContents _subpassContents;
 	MVKRenderPass* _renderPass;
+	MVKFramebuffer* _framebuffer;
 	MVKCommand* _lastMultiviewPassCmd;
 	uint32_t _renderSubpassIndex;
 	uint32_t _multiviewPassIndex;